+++ comp2013/src/main/java/comp2013/SnakeController.java
@@ -1,8 +1,21 @@
 package comp2013;
 
+import javafx.fxml.FXML;
+import javafx.scene.control.Label;
+import javafx.scene.image.ImageView;
+import javafx.scene.image.Image;
+
+
+
 public class SnakeController implements IController{
+    @FXML
+    public Label scoreLabel;
+    @FXML
+    private ImageView imageView;
     SnakeModel m_Model;
     SnakeView m_View;
+
+    public SnakeObject m_Snake;
     // Holds an instance of the controller that is used in View to set the Controller
     // This is because JavaFX creates its own instance of view before the controller
     //      can be initialised, so in order to give it a controller, it needs to be
@@ -13,12 +26,20 @@ public class SnakeController implements IController{
         m_Instance = this;
         this.m_Model = SnakeModel.getInstance();
         m_Model.setController(this);
-        this.m_View = SnakeView.getInstance();
-        m_View.setController(this);
 
-        System.out.printf("Im in the Controller!\n");
+        // Create a new snake.
+        m_Snake = new SnakeObject();
+        // Tells the model that the snake is alive.
+        m_Model.setAlive(1);
     }
 
+    /**
+     * Sets the view of the controller.
+     * @param view SnakeView to be used in the controller.
+     */
+    // I have had to do this as FXML creates its own instance of Controller,
+    //  before the view is created, so I have to set it after the fact from within view.
+    public void setView(SnakeView view){ m_View = view; }
     /**
      * Returns an instance of the Controller.
      * @return SnakeController Instance
@@ -52,4 +73,13 @@ public class SnakeController implements IController{
     public void moveSnake(int direction) {
 
     }
+
+    @FXML
+    public void initialize() {
+        Image background = new Image(getClass().getResourceAsStream("/images/UI-Background.png"));
+        imageView.setImage(background);
+
+    }
+
+
 }