+++ comp2013/src/main/java/comp2013/SnakeModel.java
@@ -48,7 +48,6 @@ public class SnakeModel implements IModel {
         return this.m_Score;
     }
 
-
     //Set the score of the game
     @Override
     public void setScore(int score) {