+++ comp2013/src/main/java/comp2013/SnakeObject.java
@@ -1,0 +1,16 @@
+package comp2013;
+
+public class SnakeObject implements Movable {
+    // Keeps track of how fast the snake is moving.
+    private int M_SnakeSpeed;
+
+    public SnakeObject(){
+        this.M_SnakeSpeed = 5;
+    }
+
+    @Override
+    public void moveSnake() {
+
+    }
+
+}