+++ comp2013/src/main/java/comp2013/SnakeThread.java
@@ -1,0 +1,22 @@
+package comp2013;
+
+import java.awt.*;
+import java.awt.Component;
+public class SnakeThread extends Thread{
+    @Override
+    public void run()
+    {
+        super.run();
+        while (true)
+        {
+            // Refresh the screen.
+            try
+            {
+                sleep(30);
+            } catch (Exception e)
+            {
+                e.printStackTrace();
+            }
+        }
+    }
+}